Fix some type errors which prevented either gcc5 or gcc6 from building; fix some warnings;

+++ src/analyzer.cpp
@@ -470,13 +470,13 @@ void analyzer::draw(int subindex, float *data, int points, bool fftdone) const
                                 lastoutL = fft_outL[_iter];
                                 //pumping up actual signal an erase surrounding
                                 // sounds
-                                fft_outL[_iter] = 0.25f * std::max(n * 0.6f * \
-                                    fabs(fft_outL[_iter]) - var1L , 1e-20f);
+                                fft_outL[_iter] = 0.25f * std::max((float)(n * 0.6f * \
+                                    fabs(fft_outL[_iter]) - var1L), 1e-20f);
                                 if(_mode == 3 or _mode == 4) {
                                     // do the same with R channel if needed
                                     lastoutR = fft_outR[_iter];
-                                    fft_outR[_iter] = 0.25f * std::max(n * \
-                                        0.6f * fabs(fft_outR[_iter]) - var1R , 1e-20f);
+                                    fft_outR[_iter] = 0.25f * std::max((float)(n * \
+                                        0.6f * fabs(fft_outR[_iter]) - var1R), 1e-20f);
                                 }
                                 break;
                         }
